python-backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.18
Django==5.2.1
django-prometheus==2.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
redis==5.0.3
googleapis-common-protos==1.59.1
grpcio==1.71.0
gunicorn==23.0.0
importlib_metadata==8.6.1
kombu==5.5.3
numpy==2.2.6
opentelemetry-api==1.33.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-django==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
pandas==2.2.3
prometheus_client==0.22.0
prompt_toolkit==3.0.51
protobuf==4.25.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
thrift==0.21.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0